receipe_master/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
beautifulsoup4
requests
mcp
google-generativeai
